requirements.txt
streamlit
pymupdf
faiss-cpu
sentence-transformers
langchain
openai